place the google preview in components
